refactor(backend): resolve #1743 — split celery_tasks.py into task domain modules
